internal cleanup (move files)
