+++ src/base/render.h
@@ -1,7 +1,7 @@
 #ifndef _RENDER_H
 #define _RENDER_H
 
-#include "vgmstream.h"
+#include "../vgmstream.h"
 
 void free_layout(VGMSTREAM* vgmstream);
 void reset_layout(VGMSTREAM* vgmstream);
