+++ src/layout/flat.c
@@ -1,6 +1,6 @@
 #include "layout.h"
 #include "../vgmstream.h"
-#include "../decode.h"
+#include "../base/decode.h"
 
 
 /* Decodes samples for flat streams.
