+++ src/libvgmstream.vcxproj
@@ -81,15 +81,15 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="api.h" />
-    <ClInclude Include="decode.h" />
-    <ClInclude Include="mixing.h" />
-    <ClInclude Include="plugins.h" />
-    <ClInclude Include="render.h" />
     <ClInclude Include="streamfile.h" />
     <ClInclude Include="streamtypes.h" />
     <ClInclude Include="util.h" />
     <ClInclude Include="vgmstream.h" />
     <ClInclude Include="vgmstream_types.h" />
+    <ClInclude Include="base\decode.h" />
+    <ClInclude Include="base\mixing.h" />
+    <ClInclude Include="base\plugins.h" />
+    <ClInclude Include="base\render.h" />
     <ClInclude Include="coding\acm_decoder_libacm.h" />
     <ClInclude Include="coding\circus_decoder_lib.h" />
     <ClInclude Include="coding\circus_decoder_lib_data.h" />
@@ -184,16 +184,16 @@
     <ClInclude Include="util\text_reader.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="decode.c" />
     <ClCompile Include="formats.c" />
-    <ClCompile Include="mixing.c" />
-    <ClCompile Include="plugins.c" />
-    <ClCompile Include="render.c" />
-    <ClCompile Include="seek.c" />
     <ClCompile Include="streamfile.c" />
     <ClCompile Include="util.c" />
     <ClCompile Include="vgmstream.c" />
+    <ClCompile Include="base\decode.c" />
     <ClCompile Include="base\info.c" />
+    <ClCompile Include="base\mixing.c" />
+    <ClCompile Include="base\plugins.c" />
+    <ClCompile Include="base\render.c" />
+    <ClCompile Include="base\seek.c" />
     <ClCompile Include="coding\acm_decoder.c" />
     <ClCompile Include="coding\acm_decoder_decode.c" />
     <ClCompile Include="coding\acm_decoder_util.c" />
