+++ src/libvgmstream.vcxproj.filters
@@ -68,18 +68,6 @@
     <ClInclude Include="api.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="decode.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="mixing.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="plugins.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="render.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="streamfile.h">
       <Filter>Header Files</Filter>
     </ClInclude>
@@ -94,6 +82,18 @@
     </ClInclude>
     <ClInclude Include="vgmstream_types.h">
       <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="base\decode.h">
+      <Filter>base\Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="base\mixing.h">
+      <Filter>base\Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="base\plugins.h">
+      <Filter>base\Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="base\render.h">
+      <Filter>base\Header Files</Filter>
     </ClInclude>
     <ClInclude Include="coding\acm_decoder_libacm.h">
       <Filter>coding\Header Files</Filter>
@@ -373,22 +373,7 @@
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="decode.c">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="formats.c">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="mixing.c">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="plugins.c">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="render.c">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="seek.c">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="streamfile.c">
@@ -400,7 +385,22 @@
     <ClCompile Include="vgmstream.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="base\decode.c">
+      <Filter>base\Source Files</Filter>
+    </ClCompile>
     <ClCompile Include="base\info.c">
+      <Filter>base\Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="base\mixing.c">
+      <Filter>base\Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="base\plugins.c">
+      <Filter>base\Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="base\render.c">
+      <Filter>base\Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="base\seek.c">
       <Filter>base\Source Files</Filter>
     </ClCompile>
     <ClCompile Include="coding\acm_decoder.c">
